backend/requirements.txt
Django==5.1.2
python-decouple==3.8
djangorestframework==3.15.2
djoser==2.2.3
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.5.0